# backend/internal/generated/go/model_create_event_request.go
/*
 * イベント管理システム API
 *
 * イベント管理システムのRESTful API - DDD/CQRSアーキテクチャに基づく実装
 *
 * API version: 1.0.0
 * Contact: dev@example.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package generated

import (
	"time"
)

type CreateEventRequest struct {

	// イベントタイトル
	Title string `json:"title"`

	// イベントの説明
	Description string `json:"description,omitempty"`

	// 会場
	Venue string `json:"venue,omitempty"`

	// 参加可能な役割
	AllowedRoles []string `json:"allowed_roles"`

	// タグ
	Tags []string `json:"tags,omitempty"`

	// 料金設定
	FeeSettings []FeeSetting `json:"fee_settings,omitempty"`

	// 日程調整タイプ
	PollType string `json:"poll_type,omitempty"`

	// 日程候補（ISO 8601形式）
	PollCandidates []time.Time `json:"poll_candidates,omitempty"`

	// 参加可能なユーザーIDのリスト（空の場合は全ユーザーが参加可能）
	AllowedUsers []string `json:"allowed_users,omitempty"`
}